Break down huge file a bit
