+++ packages/alfa-media/src/index.ts
@@ -1,1 +1,5 @@
+export * from "./feature";
 export * from "./media";
+export * from "./modifier";
+export * from "./type";
+export * from "./value";
