+++ packages/alfa-media/src/matchable.ts
@@ -1,0 +1,9 @@
+import { Device } from "@siteimprove/alfa-device";
+import { Predicate } from "@siteimprove/alfa-predicate";
+
+/**
+ * @public
+ */
+export interface Matchable {
+  readonly matches: Predicate<Device>;
+}
