+++ packages/alfa-media/src/media.ts
@@ -1,1023 +1,41 @@
-import { Comparable } from "@siteimprove/alfa-comparable";
-import {
-  Length,
-  Keyword,
-  Number,
-  Percentage,
-  Token,
-  Component,
-} from "@siteimprove/alfa-css";
+import { Token, Component } from "@siteimprove/alfa-css";
 import { Device } from "@siteimprove/alfa-device";
 import { Equatable } from "@siteimprove/alfa-equatable";
-import { Functor } from "@siteimprove/alfa-functor";
 import { Iterable } from "@siteimprove/alfa-iterable";
 import { Serializable } from "@siteimprove/alfa-json";
-import { Mapper } from "@siteimprove/alfa-mapper";
 import { Option, None } from "@siteimprove/alfa-option";
 import { Parser } from "@siteimprove/alfa-parser";
-import { Predicate } from "@siteimprove/alfa-predicate";
-import { Refinement } from "@siteimprove/alfa-refinement";
-import { Result, Err, Ok } from "@siteimprove/alfa-result";
 import { Slice } from "@siteimprove/alfa-slice";
 
 import * as json from "@siteimprove/alfa-json";
 
-import { Resolver } from "./resolver";
+import { Feature, parseMediaFeature } from "./feature";
+import type { Matchable } from "./matchable";
+import { Modifier } from "./modifier";
+import { Type } from "./type";
 
 const {
   delimited,
   either,
   end,
-  filter,
   left,
   map,
-  mapResult,
   oneOrMore,
   option,
   pair,
   right,
-  separated,
   separatedList,
   takeUntil,
   zeroOrMore,
 } = Parser;
 
-const { property, equals } = Predicate;
-
 /**
  * @public
  */
 export namespace Media {
-  /**
-   * {@link https://drafts.csswg.org/mediaqueries/#media-query-modifier}
-   */
-  export enum Modifier {
-    Only = "only",
-    Not = "not",
-  }
-
-  const parseModifier = either(
-    map(Token.parseIdent("only"), () => Modifier.Only),
-    map(Token.parseIdent("not"), () => Modifier.Not),
-  );
-
-  interface Matchable {
-    readonly matches: Predicate<Device>;
-  }
-
-  /**
-   * {@link https://drafts.csswg.org/mediaqueries/#media-type}
-   */
-  export class Type implements Matchable, Equatable, Serializable<Type.JSON> {
-    public static of(name: string): Type {
-      return new Type(name);
-    }
-
-    private readonly _name: string;
-
-    private constructor(name: string) {
-      this._name = name;
-    }
-
-    public get name(): string {
-      return this._name;
-    }
-
-    public matches(device: Device): boolean {
-      switch (this._name) {
-        case "screen":
-          return device.type === Device.Type.Screen;
-
-        case "print":
-          return device.type === Device.Type.Print;
-
-        case "speech":
-          return device.type === Device.Type.Speech;
-
-        case "all":
-          return true;
-
-        default:
-          return false;
-      }
-    }
-
-    public equals(value: unknown): value is this {
-      return value instanceof Type && value._name === this._name;
-    }
-
-    public toJSON(): Type.JSON {
-      return {
-        name: this._name,
-      };
-    }
-
-    public toString(): string {
-      return this._name;
-    }
-  }
-
-  export namespace Type {
-    export interface JSON {
-      [key: string]: json.JSON;
-      name: string;
-    }
-
-    export function isType(value: unknown): value is Type {
-      return value instanceof Type;
-    }
-  }
-
   export const { of: type, isType } = Type;
 
-  /**
-   * {@link https://drafts.csswg.org/mediaqueries/#typedef-media-type}
-   */
-  const parseType = map(
-    Token.parseIdent((ident) => {
-      switch (ident.value) {
-        // These values are not allowed as media types.
-        case "only":
-        case "not":
-        case "and":
-        case "or":
-          return false;
-
-        default:
-          return true;
-      }
-    }),
-    (ident) => Type.of(ident.value),
-  );
-
-  /**
-   * {@link https://drafts.csswg.org/mediaqueries/#media-feature}
-   */
-  export abstract class Feature<T = unknown>
-    implements
-      Matchable,
-      Iterable<Feature<T>>,
-      Equatable,
-      Serializable<Feature.JSON>
-  {
-    protected readonly _value: Option<Value<T>>;
-
-    protected constructor(value: Option<Value<T>>) {
-      this._value = value;
-    }
-
-    public abstract get name(): string;
-
-    public get value(): Option<Value<T>> {
-      return this._value;
-    }
-
-    public abstract matches(device: Device): boolean;
-
-    public equals(value: unknown): value is this {
-      return (
-        value instanceof Feature &&
-        value.name === this.name &&
-        value._value.equals(this._value)
-      );
-    }
-
-    public *iterator(): Iterator<Feature<T>> {
-      yield this;
-    }
-
-    public [Symbol.iterator](): Iterator<Feature<T>> {
-      return this.iterator();
-    }
-
-    public toJSON(): Feature.JSON {
-      return {
-        type: "feature",
-        name: this.name,
-        value: this._value.map((value) => value.toJSON()).getOr(null),
-      };
-    }
-
-    public toString(): string {
-      return `${this.name}${this._value
-        .map((value) => `: ${value}`)
-        .getOr("")}`;
-    }
-  }
-
-  export namespace Feature {
-    export interface JSON {
-      [key: string]: json.JSON;
-      type: "feature";
-      name: string;
-      value: Value.JSON | null;
-    }
-
-    export function tryFrom(
-      value: Option<Value<any>>,
-      name: string,
-    ): Result<Feature, string> {
-      switch (name) {
-        case "width":
-          return Width.tryFrom(value);
-
-        case "height":
-          return Height.tryFrom(value);
-
-        case "orientation":
-          return Orientation.tryFrom(value);
-
-        case "scripting":
-          return Scripting.tryFrom(value);
-      }
-
-      return Err.of(`Unknown media feature ${name}`);
-    }
-
-    /**
-     * {@link https://drafts.csswg.org/mediaqueries/#width}
-     */
-    class Width extends Feature<Length.Fixed> {
-      public static of(value: Value<Length.Fixed>): Width {
-        return new Width(Option.of(value));
-      }
-
-      private static _boolean = new Width(None);
-
-      public static boolean(): Width {
-        return Width._boolean;
-      }
-
-      public get name(): "width" {
-        return "width";
-      }
-
-      public matches(device: Device): boolean {
-        const {
-          viewport: { width },
-        } = device;
-
-        const value = this._value.map((value) =>
-          value.map((length) => length.resolve(Resolver.length(device))),
-        );
-
-        return width > 0
-          ? value.some((value) => value.matches(Length.of(width, "px")))
-          : value.every((value) => value.matches(Length.of(0, "px")));
-      }
-    }
-
-    namespace Width {
-      export function tryFrom(value: Option<Value>): Result<Width, string> {
-        return value
-          .map((value) =>
-            Value.Range.isRange(value) ? value.toLength() : value,
-          )
-          .map((value) => {
-            if (
-              value.hasValue(Length.isLength) &&
-              value.hasValue(
-                (value): value is Length.Fixed => !value.hasCalculation(),
-              )
-            ) {
-              return Ok.of(Width.of(value));
-            }
-
-            return Err.of(`Invalid value`);
-          })
-          .getOrElse(() => Ok.of(Width.boolean()));
-      }
-
-      export function isWidth(value: Feature): value is Width;
-
-      export function isWidth(value: unknown): value is Width;
-
-      export function isWidth(value: unknown): value is Width {
-        return value instanceof Width;
-      }
-    }
-
-    export const { isWidth } = Width;
-
-    /**
-     * {@link https://drafts.csswg.org/mediaqueries/#height}
-     */
-    class Height extends Feature<Length.Fixed> {
-      public static of(value: Value<Length.Fixed>): Height {
-        return new Height(Option.of(value));
-      }
-
-      private static _boolean = new Height(None);
-
-      public static boolean(): Height {
-        return Height._boolean;
-      }
-
-      public get name(): "height" {
-        return "height";
-      }
-
-      public matches(device: Device): boolean {
-        const {
-          viewport: { height },
-        } = device;
-
-        const value = this._value.map((value) =>
-          value.map((length) => length.resolve(Resolver.length(device))),
-        );
-
-        return height > 0
-          ? value.some((value) => value.matches(Length.of(height, "px")))
-          : value.every((value) => value.matches(Length.of(0, "px")));
-      }
-    }
-
-    namespace Height {
-      export function tryFrom(value: Option<Value>): Result<Height, string> {
-        return value
-          .map((value) =>
-            Value.Range.isRange(value) ? value.toLength() : value,
-          )
-          .map((value) => {
-            if (
-              value.hasValue(Length.isLength) &&
-              value.hasValue(
-                (value): value is Length.Fixed => !value.hasCalculation(),
-              )
-            ) {
-              return Ok.of(Height.of(value));
-            }
-
-            return Err.of(`Invalid value`);
-          })
-          .getOrElse(() => Ok.of(Height.boolean()));
-      }
-
-      export function isHeight(value: Feature): value is Height;
-
-      export function isHeight(value: unknown): value is Height;
-
-      export function isHeight(value: unknown): value is Height {
-        return value instanceof Height;
-      }
-    }
-
-    export const { isHeight } = Height;
-
-    /**
-     * {@link https://drafts.csswg.org/mediaqueries/#orientation}
-     */
-    class Orientation extends Feature<Keyword> {
-      public static of(value: Value<Keyword>): Orientation {
-        return new Orientation(Option.of(value));
-      }
-
-      private static _boolean = new Orientation(None);
-
-      public static boolean(): Orientation {
-        return Orientation._boolean;
-      }
-
-      public get name(): "orientation" {
-        return "orientation";
-      }
-
-      public matches(device: Device): boolean {
-        return this._value.every((value) =>
-          value.matches(Keyword.of(device.viewport.orientation)),
-        );
-      }
-    }
-
-    namespace Orientation {
-      export function tryFrom(
-        value: Option<Value<any>>,
-      ): Result<Orientation, string> {
-        return value
-          .map((value) => {
-            if (
-              Value.isDiscrete(value) &&
-              value.hasValue(
-                Refinement.and(
-                  Keyword.isKeyword,
-                  property("value", equals("landscape", "portrait")),
-                ),
-              )
-            ) {
-              return Ok.of(Orientation.of(value));
-            } else {
-              return Err.of(`Invalid value`);
-            }
-          })
-          .getOrElse(() => Ok.of(Orientation.boolean()));
-      }
-    }
-
-    /**
-     * {@link https://drafts.csswg.org/mediaqueries-5/#scripting}
-     */
-    class Scripting extends Feature<Keyword> {
-      public static of(value: Value<Keyword>): Scripting {
-        return new Scripting(Option.of(value));
-      }
-
-      private static _boolean = new Scripting(None);
-
-      public static boolean(): Scripting {
-        return Scripting._boolean;
-      }
-
-      public get name(): "scripting" {
-        return "scripting";
-      }
-
-      public matches(device: Device): boolean {
-        return device.scripting.enabled
-          ? this._value.every((value) => value.matches(Keyword.of("enabled")))
-          : this._value.some((value) => value.matches(Keyword.of("none")));
-      }
-    }
-
-    namespace Scripting {
-      export function tryFrom(
-        value: Option<Value<any>>,
-      ): Result<Scripting, string> {
-        return value
-          .map((value) => {
-            if (
-              Value.isDiscrete(value) &&
-              value.hasValue(
-                Refinement.and(
-                  Keyword.isKeyword,
-                  property("value", equals("none", "enabled", "initial-only")),
-                ),
-              )
-            ) {
-              return Ok.of(Scripting.of(value));
-            } else {
-              return Err.of(`Invalid value`);
-            }
-          })
-          .getOrElse(() => Ok.of(Scripting.boolean()));
-      }
-    }
-
-    export function isFeature(value: unknown): value is Feature {
-      return value instanceof Feature;
-    }
-  }
-
   export const { isFeature } = Feature;
-
-  export enum Comparison {
-    LessThan = "<",
-    LessThanOrEqual = "<=",
-    Equal = "=",
-    GreaterThan = ">",
-    GreaterThanOrEqual = ">=",
-  }
-
-  /**
-   * {@link https://drafts.csswg.org/mediaqueries/#typedef-mf-name}
-   */
-  const parseFeatureName = map(Token.parseIdent(), (ident) =>
-    ident.value.toLowerCase(),
-  );
-
-  /**
-   * {@link https://drafts.csswg.org/mediaqueries/#typedef-mf-value}
-   *
-   * @remarks
-   * We currently do not support calculations in media queries
-   */
-  const parseFeatureValue = either(
-    either(
-      filter(
-        Number.parse,
-        (number) => !number.hasCalculation(),
-        () => "Calculations no supported in media queries",
-      ),
-      map(Token.parseIdent(), (ident) => Keyword.of(ident.value.toLowerCase())),
-    ),
-    either(
-      map(
-        separated(
-          Token.parseNumber((number) => number.isInteger),
-          delimited(option(Token.parseWhitespace), Token.parseDelim("/")),
-          Token.parseNumber((number) => number.isInteger),
-        ),
-        ([left, right]) => Percentage.of(left.value / right.value),
-      ),
-      filter(
-        Length.parse,
-        (length) => !length.hasCalculation(),
-        () => "Calculations no supported in media queries",
-      ),
-    ),
-  );
-
-  /**
-   * {@link https://drafts.csswg.org/mediaqueries/#typedef-mf-plain}
-   */
-  const parseFeaturePlain = mapResult(
-    separated(
-      parseFeatureName,
-      delimited(option(Token.parseWhitespace), Token.parseColon),
-      parseFeatureValue,
-    ),
-    ([name, value]) => {
-      if (name.startsWith("min-") || name.startsWith("max-")) {
-        const range = name.startsWith("min-")
-          ? Value.minimumRange
-          : Value.maximumRange;
-
-        name = name.slice(4);
-
-        return Feature.tryFrom(
-          Option.of(range(Value.bound(value, /* isInclusive */ true))),
-          name,
-        );
-      } else {
-        return Feature.tryFrom(Option.of(Value.discrete(value)), name);
-      }
-    },
-  );
-
-  /**
-   * {@link https://drafts.csswg.org/mediaqueries/#typedef-mf-boolean}
-   */
-  const parseFeatureBoolean = mapResult(parseFeatureName, (name) =>
-    Feature.tryFrom(None, name),
-  );
-
-  /**
-   * {@link https://drafts.csswg.org/mediaqueries/#typedef-mf-lt}
-   */
-  const parseFeatureLessThan = map(
-    right(Token.parseDelim("<"), option(Token.parseDelim("="))),
-    (equal) =>
-      equal.isNone() ? Comparison.LessThan : Comparison.LessThanOrEqual,
-  );
-
-  /**
-   * {@link https://drafts.csswg.org/mediaqueries/#typedef-mf-gt}
-   */
-  const parseFeatureGreaterThan = map(
-    right(Token.parseDelim(">"), option(Token.parseDelim("="))),
-    (equal) =>
-      equal.isNone() ? Comparison.GreaterThan : Comparison.GreaterThanOrEqual,
-  );
-
-  /**
-   * {@link https://drafts.csswg.org/mediaqueries/#typedef-mf-eq}
-   */
-  const parseFeatureEqual = map(Token.parseDelim("="), () => Comparison.Equal);
-
-  /**
-   * {@link https://drafts.csswg.org/mediaqueries/#typedef-mf-comparison}
-   */
-  const parseFeatureComparison = either(
-    parseFeatureEqual,
-    parseFeatureLessThan,
-    parseFeatureGreaterThan,
-  );
-
-  /**
-   * {@link https://drafts.csswg.org/mediaqueries/#typedef-mf-range}
-   */
-  const parseFeatureRange = either(
-    // <mf-value> <mf-lt> <mf-name> <mf-lt> <mf-value>
-    mapResult(
-      pair(
-        map(
-          pair(
-            parseFeatureValue,
-            delimited(option(Token.parseWhitespace), parseFeatureLessThan),
-          ),
-          ([value, comparison]) =>
-            Value.bound(
-              value,
-              /* isInclusive */ comparison === Comparison.LessThanOrEqual,
-            ),
-        ),
-        pair(
-          delimited(option(Token.parseWhitespace), parseFeatureName),
-          map(
-            pair(
-              delimited(option(Token.parseWhitespace), parseFeatureLessThan),
-              parseFeatureValue,
-            ),
-            ([comparison, value]) =>
-              Value.bound(
-                value,
-                /* isInclusive */ comparison === Comparison.LessThanOrEqual,
-              ),
-          ),
-        ),
-      ),
-      ([minimum, [name, maximum]]) =>
-        Feature.tryFrom(Option.of(Value.range(minimum, maximum)), name),
-    ),
-
-    // <mf-value> <mf-gt> <mf-name> <mf-gt> <mf-value>
-    mapResult(
-      pair(
-        map(
-          pair(
-            parseFeatureValue,
-            delimited(option(Token.parseWhitespace), parseFeatureGreaterThan),
-          ),
-          ([value, comparison]) =>
-            Value.bound(
-              value,
-              /* isInclusive */ comparison === Comparison.GreaterThanOrEqual,
-            ),
-        ),
-        pair(
-          delimited(option(Token.parseWhitespace), parseFeatureName),
-          map(
-            pair(
-              delimited(option(Token.parseWhitespace), parseFeatureGreaterThan),
-              parseFeatureValue,
-            ),
-            ([comparison, value]) =>
-              Value.bound(
-                value,
-                /* isInclusive */ comparison === Comparison.GreaterThanOrEqual,
-              ),
-          ),
-        ),
-      ),
-      ([maximum, [name, minimum]]) =>
-        Feature.tryFrom(Option.of(Value.range(minimum, maximum)), name),
-    ),
-
-    // <mf-name> <mf-comparison> <mf-value>
-    mapResult(
-      pair(
-        parseFeatureName,
-        pair(
-          delimited(option(Token.parseWhitespace), parseFeatureComparison),
-          parseFeatureValue,
-        ),
-      ),
-      ([name, [comparison, value]]) => {
-        switch (comparison) {
-          case Comparison.Equal:
-            return Feature.tryFrom(
-              Option.of(
-                Value.range(
-                  Value.bound(value, /* isInclude */ true),
-                  Value.bound(value, /* isInclude */ true),
-                ),
-              ),
-              name,
-            );
-
-          case Comparison.LessThan:
-          case Comparison.LessThanOrEqual:
-            return Feature.tryFrom(
-              Option.of(
-                Value.maximumRange(
-                  Value.bound(
-                    value,
-                    /* isInclusive */ comparison === Comparison.LessThanOrEqual,
-                  ),
-                ),
-              ),
-              name,
-            );
-
-          case Comparison.GreaterThan:
-          case Comparison.GreaterThanOrEqual:
-            return Feature.tryFrom(
-              Option.of(
-                Value.minimumRange(
-                  Value.bound(
-                    value,
-                    /* isInclusive */ comparison ===
-                      Comparison.GreaterThanOrEqual,
-                  ),
-                ),
-              ),
-              name,
-            );
-        }
-      },
-    ),
-
-    // <mf-value> <mf-comparison> <mf-name>
-    mapResult(
-      pair(
-        parseFeatureValue,
-        pair(
-          delimited(option(Token.parseWhitespace), parseFeatureComparison),
-          parseFeatureName,
-        ),
-      ),
-      ([value, [comparison, name]]) => {
-        switch (comparison) {
-          case Comparison.Equal:
-            return Feature.tryFrom(
-              Option.of(
-                Value.range(
-                  Value.bound(value, /* isInclude */ true),
-                  Value.bound(value, /* isInclude */ true),
-                ),
-              ),
-              name,
-            );
-
-          case Comparison.LessThan:
-          case Comparison.LessThanOrEqual:
-            return Feature.tryFrom(
-              Option.of(
-                Value.minimumRange(
-                  Value.bound(
-                    value,
-                    /* isInclusive */ comparison === Comparison.LessThanOrEqual,
-                  ),
-                ),
-              ),
-              name,
-            );
-
-          case Comparison.GreaterThan:
-          case Comparison.GreaterThanOrEqual:
-            return Feature.tryFrom(
-              Option.of(
-                Value.maximumRange(
-                  Value.bound(
-                    value,
-                    /* isInclusive */ comparison ===
-                      Comparison.GreaterThanOrEqual,
-                  ),
-                ),
-              ),
-              name,
-            );
-        }
-      },
-    ),
-  );
-
-  /**
-   * {@link https://drafts.csswg.org/mediaqueries/#typedef-media-feature}
-   */
-  const parseFeature = delimited(
-    Token.parseOpenParenthesis,
-    delimited(
-      option(Token.parseWhitespace),
-      either(parseFeatureRange, parseFeaturePlain, parseFeatureBoolean),
-    ),
-    Token.parseCloseParenthesis,
-  );
-
-  export interface Value<T = unknown>
-    extends Functor<T>,
-      Serializable<Value.JSON> {
-    map<U>(mapper: Mapper<T, U>): Value<U>;
-    matches(value: T): boolean;
-    hasValue<U extends T>(refinement: Refinement<T, U>): this is Value<U>;
-    toJSON(): Value.JSON;
-  }
-
-  export namespace Value {
-    export interface JSON {
-      [key: string]: json.JSON;
-      type: string;
-    }
-
-    export class Discrete<T = unknown>
-      implements Value<T>, Serializable<Discrete.JSON<T>>
-    {
-      public static of<T>(value: T): Discrete<T> {
-        return new Discrete(value);
-      }
-
-      private readonly _value: T;
-
-      private constructor(value: T) {
-        this._value = value;
-      }
-
-      public get value(): T {
-        return this._value;
-      }
-
-      public map<U>(mapper: Mapper<T, U>): Discrete<U> {
-        return new Discrete(mapper(this._value));
-      }
-
-      public matches(value: T): boolean {
-        return Equatable.equals(this._value, value);
-      }
-
-      public hasValue<U extends T>(
-        refinement: Refinement<T, U>,
-      ): this is Discrete<U> {
-        return refinement(this._value);
-      }
-
-      public toJSON(): Discrete.JSON<T> {
-        return {
-          type: "discrete",
-          value: Serializable.toJSON(this._value),
-        };
-      }
-    }
-
-    export namespace Discrete {
-      export interface JSON<T> {
-        [key: string]: json.JSON;
-        type: "discrete";
-        value: Serializable.ToJSON<T>;
-      }
-
-      export function isDiscrete<T>(value: unknown): value is Discrete<T> {
-        return value instanceof Discrete;
-      }
-    }
-
-    export const { of: discrete, isDiscrete } = Discrete;
-
-    export class Range<T = unknown>
-      implements Value<T>, Serializable<Range.JSON<T>>
-    {
-      public static of<T>(minimum: Bound<T>, maximum: Bound<T>): Range<T> {
-        return new Range(Option.of(minimum), Option.of(maximum));
-      }
-
-      public static minimum<T>(minimum: Bound<T>): Range<T> {
-        return new Range(Option.of(minimum), None);
-      }
-
-      public static maximum<T>(maximum: Bound<T>): Range<T> {
-        return new Range(None, Option.of(maximum));
-      }
-
-      private readonly _minimum: Option<Bound<T>>;
-      private readonly _maximum: Option<Bound<T>>;
-
-      private constructor(
-        minimum: Option<Bound<T>>,
-        maximum: Option<Bound<T>>,
-      ) {
-        this._minimum = minimum;
-        this._maximum = maximum;
-      }
-
-      public get minimum(): Option<Bound<T>> {
-        return this._minimum;
-      }
-
-      public get maximum(): Option<Bound<T>> {
-        return this._maximum;
-      }
-
-      public map<U>(mapper: Mapper<T, U>): Range<U> {
-        return new Range(
-          this._minimum.map((bound) => bound.map(mapper)),
-          this._maximum.map((bound) => bound.map(mapper)),
-        );
-      }
-
-      public toLength(): Range<T | Length<"px">> {
-        return this.map((bound) =>
-          Refinement.and(
-            Number.isNumber,
-            (value) => !value.hasCalculation() && value.value === 0,
-          )(bound)
-            ? Length.of(0, "px")
-            : bound,
-        );
-      }
-
-      public matches(value: T): boolean {
-        if (!Comparable.isComparable<T>(value)) {
-          return false;
-        }
-
-        // Since we need to match both bounds, we return false if one is not
-        // matched and keep true for the default return at the end.
-        for (const minimum of this._minimum) {
-          if (minimum.isInclusive) {
-            // value is inclusively larger than the minimum if it is not
-            // strictly smaller than it.
-            if (value.compare(minimum.value) < 0) {
-              return false;
-            }
-          } else {
-            if (value.compare(minimum.value) <= 0) {
-              return false;
-            }
-          }
-        }
-
-        for (const maximum of this._maximum) {
-          if (maximum.isInclusive) {
-            if (value.compare(maximum.value) > 0) {
-              return false;
-            }
-          } else {
-            if (value.compare(maximum.value) >= 0) {
-              return false;
-            }
-          }
-        }
-
-        return true;
-      }
-
-      public hasValue<U extends T>(
-        refinement: Refinement<T, U>,
-      ): this is Discrete<U> {
-        return (
-          this._minimum.every((bound) => refinement(bound.value)) &&
-          this._maximum.every((bound) => refinement(bound.value))
-        );
-      }
-
-      public toJSON(): Range.JSON<T> {
-        return {
-          type: "range",
-          minimum: this._minimum.map((bound) => bound.toJSON()).getOr(null),
-          maximum: this._maximum.map((bound) => bound.toJSON()).getOr(null),
-        };
-      }
-    }
-
-    export namespace Range {
-      export interface JSON<T> {
-        [key: string]: json.JSON;
-        type: "range";
-        minimum: Bound.JSON<T> | null;
-        maximum: Bound.JSON<T> | null;
-      }
-
-      export function isRange<T>(value: unknown): value is Range<T> {
-        return value instanceof Range;
-      }
-    }
-
-    export const {
-      of: range,
-      minimum: minimumRange,
-      maximum: maximumRange,
-      isRange,
-    } = Range;
-
-    export class Bound<T = unknown>
-      implements Functor<T>, Serializable<Bound.JSON<T>>
-    {
-      public static of<T>(value: T, isInclusive: boolean): Bound<T> {
-        return new Bound(value, isInclusive);
-      }
-
-      private readonly _value: T;
-      private readonly _isInclusive: boolean;
-
-      private constructor(value: T, isInclusive: boolean) {
-        this._value = value;
-        this._isInclusive = isInclusive;
-      }
-
-      public get value(): T {
-        return this._value;
-      }
-
-      public get isInclusive(): boolean {
-        return this._isInclusive;
-      }
-
-      public map<U>(mapper: Mapper<T, U>): Bound<U> {
-        return new Bound(mapper(this._value), this._isInclusive);
-      }
-
-      public hasValue<U extends T>(
-        refinement: Refinement<T, U>,
-      ): this is Bound<U> {
-        return refinement(this._value);
-      }
-
-      public toJSON(): Bound.JSON<T> {
-        return {
-          value: Serializable.toJSON(this._value),
-          isInclusive: this._isInclusive,
-        };
-      }
-    }
-
-    export namespace Bound {
-      export interface JSON<T> {
-        [key: string]: json.JSON;
-        value: Serializable.ToJSON<T>;
-        isInclusive: boolean;
-      }
-    }
-
-    export const { of: bound } = Bound;
-  }
 
   export class And
     implements Matchable, Iterable<Feature>, Equatable, Serializable<And.JSON>
@@ -1263,7 +281,7 @@ export namespace Media {
       ),
       Token.parseCloseParenthesis,
     ),
-    parseFeature,
+    parseMediaFeature,
   );
 
   /**
@@ -1440,8 +458,8 @@ export namespace Media {
       map(
         pair(
           pair(
-            option(delimited(option(Token.parseWhitespace), parseModifier)),
-            parseType,
+            option(delimited(option(Token.parseWhitespace), Modifier.parse)),
+            Type.parse,
           ),
           option(
             right(
