+++ packages/alfa-media/src/value/index.ts
@@ -1,0 +1,1 @@
+export * from "./value";
